Update image template and vagrantfile
The base image source URLs referenced in the the previous version
of template.json are no longer available. The relevant template
parameters have been updated to URLs that referencew the latest
version of the base image.

The config option removed from the Vagrantfile is no longer valid
and was removed completely.